Remove selection relative code from main controller
