+++ geoportailv3/templates/index.html
@@ -211,7 +211,7 @@
           <app-draw app-draw-map="::mainCtrl.map"
             app-draw-active="mainCtrl.drawOpen"
             app-draw-features="::mainCtrl.drawnFeatures"
-            app-draw-selectinteraction="::mainCtrl.selectInteraction">
+            app-draw-selectedfeatures="::mainCtrl.selectedFeatures">
           </app-draw>
         </li>
         <li class="hidden-xs measure icon" ng-class="mainCtrl.measureOpen ? 'active' : ''">
